[chore] #1 home_page 파일 분리
